Ajout d'une superclasse Utilisateur
