+++ app/src/main/java/ca/uottawa/eecs/seg2505/benevolapp/model/Utilisateur.java
@@ -1,0 +1,40 @@
+package ca.uottawa.eecs.seg2505.benevolapp.model;
+
+public abstract class Utilisateur {
+    protected String codePostal;
+    protected String ville;
+    protected String numeroTelephone;
+    protected String courriel;
+
+    public String getCodePostal() {
+        return codePostal;
+    }
+
+    public void setCodePostal(String codePostal) {
+        this.codePostal = codePostal;
+    }
+
+    public String getVille() {
+        return ville;
+    }
+
+    public void setVille(String ville) {
+        this.ville = ville;
+    }
+
+    public String getNumeroTelephone() {
+        return numeroTelephone;
+    }
+
+    public void setNumeroTelephone(String numeroTelephone) {
+        this.numeroTelephone = numeroTelephone;
+    }
+
+    public String getCourriel() {
+        return courriel;
+    }
+
+    public void setCourriel(String courriel) {
+        this.courriel = courriel;
+    }
+}
